Shifted delete_article functionality to a button using 'Post' method and added if statement ensuring that deletion will only occur via the button.

+++ Code/Blog/delete_article.php
@@ -4,37 +4,41 @@ require("includes/database.php");
 require("includes/article.php");
 require("includes/url.php");
 
-$conn = getDatabase();
-if (isset($_GET["id"]) && is_numeric($_GET['id'])) {
-	$id = $_GET['id'];
-	$article = get_article($conn, $id);
-	if($article){
-		$title = $article['title'];
-		$content = $article['content'];
-		$published_at = $article['published_at'];
+if($_SERVER["REQUEST_METHOD"] == "POST") {
+	$conn = getDatabase();
+	if (isset($_GET["id"]) && is_numeric($_GET['id'])) {
+		$id = $_GET['id'];
+		$article = get_article($conn, $id);
+		if($article){
+			$title = $article['title'];
+			$content = $article['content'];
+			$published_at = $article['published_at'];
+		} else {
+			die("Article not found, buddy boy");
+		}
 	} else {
-		die("Article not found, buddy boy");
+		$article = null;
+		die("ID not provided");
+	}
+
+	// using a Prepared Statement to foil SQL injection attacks.
+	$sql_query = "DELETE FROM articles WHERE id = ?";
+
+	$prep_stmt = mysqli_prepare($conn, $sql_query);
+
+	if($prep_stmt === false){
+		echo mysqli_error($conn);
+	} else {
+		// binding params to create prepared statement
+		mysqli_stmt_bind_param($prep_stmt, "i", $id);
+
+		// executing prepared statement
+		if(mysqli_stmt_execute($prep_stmt)){
+			redirect("/Code/Blog/index.php");
+		} else {
+			echo mysqli_stmt_error($prep_stmt);
+		}
 	}
 } else {
-	$article = null;
-	die("ID not provided");
-}
-
-// using a Prepared Statement to foil SQL injection attacks.
-$sql_query = "DELETE FROM articles WHERE id = ?";
-
-$prep_stmt = mysqli_prepare($conn, $sql_query);
-
-if($prep_stmt === false){
-	echo mysqli_error($conn);
-} else {
-	// binding params to create prepared statement
-	mysqli_stmt_bind_param($prep_stmt, "i", $id);
-
-	// executing prepared statement
-	if(mysqli_stmt_execute($prep_stmt)){
-		redirect("/Code/Blog/index.php");
-	} else {
-		echo mysqli_stmt_error($prep_stmt);
-	}
+	redirect("/Code/Blog/index.php");
 }
